Small cleanup
