+++ src/template-engine/files/report-handlers/CsRxHandler.ts
@@ -1,7 +1,15 @@
 import { WorkBook } from "xlsx";
 import { ReportSheets } from "../ReportFile";
 import { BaseReportHandler } from "./BaseHandler";
-import { csrxSheet } from "../../../utils/sheets";
+
+type csrxSheet = {
+  'Drug Name': string;
+  Family: string;
+  'mg/day': number;
+  'DEA#': string;
+  Qty: number;
+  'DEA Sched': string;
+}
 
 export class CsRxHandler extends BaseReportHandler<csrxSheet> {
